requirements.txt
py-cord
python-dotenv
requests
datetime